Fix CI: Publish parser to mavenLocal before building plugin
The jafar-gradle-plugin depends on jafar-parser, but when building
in separate gradle invocations, the plugin can't find the parser
(it hasn't been published to Maven Central yet during the build).

Solution: Add publishToMavenLocal to the first gradle command so
the parser is available in mavenLocal (which the plugin's repos
already includes) when building the plugin.